requirements.txt
alembic==1.11.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
asyncpg==0.27.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cohere==4.37
comm==0.1.3
databases==0.7.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
environs==9.5.0
executing==1.2.0
fastapi==0.97.0
FastAPI-SQLAlchemy==0.2.1
fastjsonschema==2.17.1
fqdn==1.5.1
grpcio==1.53.0
h11==0.14.0
httptools==0.5.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter-lsp==2.1.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
mistune==2.0.5
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook_shim==0.2.3
numpy==1.24.3
overrides==7.3.1
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pip==23.0.1
platformdirs==3.5.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.1
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
Pygments==2.15.1
pymilvus==2.2.8
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
Send2Trash==1.8.2
setuptools==66.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.2
tzdata==2023.3
ujson==5.7.0
uri-template==1.2.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
wheel==0.38.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
retry==0.9.2
configparser==6.0.0
annoy==1.17.3
langchain-core==0.1.42
langchain==0.1.16
tiktoken==0.6.0
openai==1.17.1
flashrank==0.2.0
sqlalchemy==2.0.29